src/coordinator/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.3
prometheus_client
psutil